+++ protected/humhub/components/Module.php
@@ -48,7 +48,7 @@ class Module extends \yii\base\Module
 
         // Set settings component
         $this->set('settings', [
-            'class' => SettingsManager::className(),
+            'class' => SettingsManager::class,
             'moduleId' => $this->id
         ]);
     }