+++ protected/humhub/docs/CHANGELOG.md
@@ -7,6 +7,7 @@ HumHub Change Log
 ---------------------------
 
 - Fix #3359: Weekly summary e-mails are not sent in default configuration
+- Fix: Space module permission configuration of non installed space modules visible in permission grid
 
 
 1.3.7  (October 23, 2018)